# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MyZubster Gateway API",
    "version": "1.0.0",
    "description": "Complete API documentation for MyZubster Gateway"
  },
  "servers": [
    {
      "url": "http://localhost:10000",
      "description": "Local"
    },
    {
      "url": "https://myzubster-gateway.onrender.com",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/buy-myz": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create MYZ purchase order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userTariWallet",
                  "amountMYZ"
                ],
                "properties": {
                  "userTariWallet": {
                    "type": "string"
                  },
                  "amountMYZ": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/escrow/create": {
      "post": {
        "tags": [
          "Escrow"
        ],
        "summary": "Create general escrow",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "escrowId",
                  "buyer",
                  "seller",
                  "amount"
                ],
                "properties": {
                  "escrowId": {
                    "type": "string"
                  },
                  "buyer": {
                    "type": "string"
                  },
                  "seller": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/stake/stake": {
      "post": {
        "tags": [
          "Stake"
        ],
        "summary": "Stake tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "amount",
                  "blocks"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "blocks": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/robot/escrow/create": {
      "post": {
        "tags": [
          "Robot Escrow"
        ],
        "summary": "Create robot job escrow",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "jobId",
                  "clientId",
                  "robotId",
                  "amount",
                  "currency"
                ],
                "properties": {
                  "jobId": {
                    "type": "string"
                  },
                  "clientId": {
                    "type": "string"
                  },
                  "robotId": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "currency": {
                    "type": "string",
                    "enum": [
                      "MYZ",
                      "XMR"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  }
}